+++ docs/source/changes.rst
@@ -11,10 +11,10 @@ v0.2.0
 
 **Descriptor Generator**:
 
-* Add ``xenonpy.descriptor.Fingerprint`` descriptor generator. (`Issues #21 <https://github.com/yoshida-lab/XenonPy/issues/21>`_)
-* Add ``xenonpy.descriptor.OrbitalFieldMatrix`` descriptor generator. (`Issues #22 <https://github.com/yoshida-lab/XenonPy/issues/22>`_)
+* Added ``xenonpy.descriptor.Fingerprint`` descriptor generator. (`Issues #21 <https://github.com/yoshida-lab/XenonPy/issues/21>`_)
+* Added ``xenonpy.descriptor.OrbitalFieldMatrix`` descriptor generator. (`Issues #22 <https://github.com/yoshida-lab/XenonPy/issues/22>`_)
 
 
 **API Changes**:
 
-* Allow ``BaseDescriptor`` class to use anonymous/renamed input. (`Issues #10 <https://github.com/yoshida-lab/XenonPy/issues/10>`_)
+* Allowed ``BaseDescriptor`` class to use anonymous/renamed input. (`Issues #10 <https://github.com/yoshida-lab/XenonPy/issues/10>`_)